amf_ue_ngap_id type changes changes

+++ lte/gateway/c/oai/lib/3gpp/3gpp_38.413.h
@@ -67,7 +67,7 @@ typedef struct ngap_ue_aggregate_maximum_bit_rate_s {
   bit_rate_t ul;
 } ngap_ue_aggregate_maximum_bit_rate_t;
 
-typedef uint32_t amf_ue_ngap_id_t;
+typedef uint64_t amf_ue_ngap_id_t;
 typedef uint32_t ran_ue_ngap_id_t;
 
 typedef struct pdusession_setup_item_s {
